Exception handling - delete

+++ src/main/java/com/joao_roberto/course/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.joao_roberto.course.entities.User;
 import com.joao_roberto.course.repositories.UserRepository;
+import com.joao_roberto.course.services.exceptions.DatabaseException;
 import com.joao_roberto.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -35,7 +37,18 @@ public class UserService {
 	}
 	
 	public void delete(Long id) {
-		userRepository.deleteById(id);
+		try {
+			if(userRepository.existsById(id)) {
+				userRepository.deleteById(id);		
+			}
+			else {
+				throw new ResourceNotFoundException(id);
+			}
+		}
+		catch(DataIntegrityViolationException e) {
+		    throw new DatabaseException(e.getMessage());	
+		}
+		
 	}
 	
 	public User update(Long id, User obj) {
